backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
psycopg2-binary
finnhub-python
yfinance
feedparser
transformers
torch
APScheduler
requests
python-dotenv
beautifulsoup4
scikit-learn
numpy
python-multipart
